+++ src/intelligence_layer/evaluation/evaluation/evaluator.py
@@ -222,7 +222,10 @@ class Evaluator(Generic[Input, Output, ExpectedOutput, Evaluation]):
 
     @final
     def evaluate_runs(
-        self, *run_ids: str, num_examples: Optional[int] = None
+        self,
+        *run_ids: str,
+        num_examples: Optional[int] = None,
+        abort_on_error: bool = False,
     ) -> EvaluationOverview:
         """Evaluates all generated outputs in the run.
 
@@ -239,6 +242,7 @@ class Evaluator(Generic[Input, Output, ExpectedOutput, Evaluation]):
                 specific evaluation. The method compares all run of the provided ids to each other.
             num_examples: The number of examples which should be evaluated from the given runs.
                 Always the first n runs stored in the evaluation repository
+            abort_on_error: Flag to abort all evaluations when an error occurs. Defaults to False.
 
         Returns:
             EvaluationOverview: An overview of the evaluation. Individual :class:`Evaluation`s will not be
@@ -339,7 +343,7 @@ class Evaluator(Generic[Input, Output, ExpectedOutput, Evaluation]):
             ],
         ) -> None:
             example, eval_id, example_outputs = args
-            self.evaluate(example, eval_id, *example_outputs)
+            self.evaluate(example, eval_id, abort_on_error, *example_outputs)
 
         with ThreadPoolExecutor(max_workers=10) as executor:
             tqdm(
@@ -362,6 +366,7 @@ class Evaluator(Generic[Input, Output, ExpectedOutput, Evaluation]):
         self,
         example: Example[Input, ExpectedOutput],
         evaluation_id: str,
+        abort_on_error: bool,
         *example_outputs: SuccessfulExampleOutput[Output],
     ) -> None:
         try:
@@ -372,6 +377,11 @@ class Evaluator(Generic[Input, Output, ExpectedOutput, Evaluation]):
                 )
             )
         except Exception as e:
+            if abort_on_error:
+                raise e
+            print(
+                f'FAILED EVALUATION: example {example.id}, {type(e).__qualname__}: "{e}"'
+            )
             result = FailedExampleEvaluation.from_exception(e)
         self._evaluation_repository.store_example_evaluation(
             ExampleEvaluation(